Extract HIR map intrinsic helpers
Move map lookup, keys, contains, and default intrinsic type/ownership lowering into hir/maps.rs so the HIR facade continues shrinking behind the compiler.hir boundary.

Update the compiler source decomposition plan and ratchet ceilings for hir.rs at 5953 lines, top-file share 0.8341, top-file lines 74142, and the new maps.rs helper module.

Validation: cargo fmt --manifest-path stage1/crates/axiomc/Cargo.toml; python3 scripts/ci/report-compiler-source-monoliths.py --json --check-plan --check-ratchet; make stage1-hir-boundary; make stage1-compiler-source-monoliths-test; make stage1-compiler-source-monoliths; cargo test --manifest-path stage1/Cargo.toml -p axiomc --lib hir::boundary_tests -- --nocapture; cargo test --manifest-path stage1/Cargo.toml -p axiomc --lib; git diff --check; git diff --cached --check.
